// src/models/asistencia.model.ts
export interface IAsistencia {
  ASISTENCIA_ID: any;
  ASISTENCIA_EMPLEADOID: string;
  ASISTENCIA_FECHA: any;
  ASISTENCIA_HORA: any;
  ASISTENCIA_FECHAC: any;
  ASISTENCIA_HORAC: any;
  ASISTENCIA_IDPUESTO: any;
  COPTO_DESCRIPCION: any;
  ASISTENCIA_CARGOID: any;
  CLIE_NOMBRE: any;
  CLIUBIC_NOMBRE: any;
  CLIUBIC_DIRECCION: any;
  HORA_FECINI: any;
  HORA_HORAINI: any;
  HORA_FECFIN: any;
  HORA_HORAFIN: any;
  HORA_INCIDENCIA: any;
  ESTADO_ASISTENCIA: any;
}